+++ testsuite/tests/printer/AnnotationNoListTuplePuns.hs
@@ -1,0 +1,17 @@
+{-# language NoListTuplePuns #-}
+module AnnotationNoListTuplePuns where
+
+type A =
+  -- comment pre
+  [
+    -- comment inside
+  ]
+  -- comment post
+
+type B =
+  -- comment pre
+  [
+    -- comment inside
+    Bool
+  ]
+  -- comment post